New parent() method and broadcasting.
util.event() is basically a strange way to implement code sharing across
different objects, with event.emitter() triggering object creation.  It lets
you mix and match type behaviors, and serves as a sort of "object factory."
Listeners are meant to be defined in a more static way--when the web app is
being initialized.

This isn't a suitable way to allow widgets to communicate with each other
in a loosely decoupled way, as the name "event" might apply.

Enter broadcast().
* Use Widget.broadcast() to send messages out to registered listeners. Specify
  types to filter the listeners. The types mean the same thing they do for
  events.
* Use this.register() to allow an object to listen. This is made explicit so
  that we don't needlessly track widgets that don't need to be tracked.