New debug mode for redoak server.
Although not quite mature, redoak's basic architecture is stable enough to need
some great tools. This is the first step towards that.

You can now navigate to localhost:3000/debug/ in order to see the entire redoak
data structure in memory, including the file data and the tree itself. The
JSON viewer is rough but functional.

Some potential future use cases:
* Examine files and their dependencies.
* Look at how redoak parsed the templates, with the ability to navigate to
  individual widget pages.
  * Examine the autogenerated JS/CSS/HTML.
  * See warnings or suggestions.
  * Possibly autogenerate your own JS.
  * A sandbox for using widgets.