Legwork for own template parser.
redoak has been using mustache to do its parsing. It was a great place to
start, but has become not ideal due to redoak's partial support for it in
server-side template rendering.

redoak currently has a hack to know all the unbound variables for a template,
which it needs when it figures out what DOM elements need to be modified when
a template changes. With an AST, this gets much easier.

Additionally, this will eventually remove the runtime mustache dependency by
generating the final innerHTML string in the autogenerated code.

This is the first step towards a smarter redoak that generates smaller, smarter
code.