Fix tests and remove mustache completely.
Mustache is now no longer used by the server code, and this uncovered many
bugs. The mustache derivative is being called "oakstache" for the time being.

It's now possible to bring loops and inversions to the generated JS. We could
even start generating code for lists of widgets, though there's some tough
work there for updating an existing widget.