Legwork for walker.
After spending time thinking about how redoak can be used outside of a node
stack, I started playing with a declarative mini-language. The language
specifies a walker that allows you to reduce JSON to a string.  Like CSS, rules
tell the walker how to render the JSON.

The walker can be used with redoak. By making a POST to the development engine
or running redoak via the command line, a user can generate server side
templates for their specific stack, which she can then use to render widgets
or the root HTML on their server, if they so wish.

This idea seems powerful to me, but I'm worried it will be confusing. I have
no idea if this idea will stick.