Reorg, walker fun page, and walker used for htmljs
Lots to cover over the past few days:
* walker is starting to work beautifully, and is covered by a decent amount of
  tests at this point. We now use it for generating the html JS, and I suspect
  it has a broader set of applications than just that in redoak.
* walker has a toy page you can get to with `redoak widgets/walker.html`.
  I used this page to instantaneously debug walker. I continue to be convinced
  that this is the way to program all the things.
* lib was looking a little hefty, so I moved the gentests and reftests out to
  root. I also put some redoak generating templates into templates/ and
  moved jsonviewer to the new widgets/ directory.

walker seems mature enough to use for its original purpose: as a way to
generate templates

The next big step is to get the walker toy onto debug. This will let devs write
programs directly on /debug to get a feel for the walker language. It could
include a mustache example.