New use syntax (tests still broken)
Specifying JSON in a use tag is no longer permitted. Instead, you may specify
objects recursively using the "obj" tag, which will add elements to the
"children" key of the JSON object.

With this update, we lose the ability to specify objects recursively except
specifically through a "children" key. In return, we get a syntax that works
better with common use cases that need to include many lines of HTML.

For example, a blog post app might now look like:

<template mx='blog'>
  <div class='entries'>
    {{#children}}
      <use mx='blog.item'></use>
    {{/children}}
  </div>
</template>

<template mx='blog.item'>...

<use mx='blog'>
  <obj date='Dec 25 2012'>
    <p>It's Christmas! ...</p>
    <p>Lorem ipsum...</p>
  </obj>
  ...
</use>

Before, we'd have to specify an awkward JSON representation of this object.

This new syntax also requires a smarter use tag, which I am working on next.