Fixed oakstache and walker unit tests!
Finally managed to get through all the unit test problems. :) We're almost
ready to convert everything to the new format and take advantage of the new,
simplified template parser.

Thanks to inspiration from an old Aaron Swartz comment, I've decided to change
'{.if blah}...{/}' to '{if blah}...{/}'. The dot is just extra typing. I had
thought there might be conflicts with if and loop and template vars, but since
identifiers with spaces aren't allowed anyways, there isn't any conflict.

I think I'd like to change the template to use '{var}' as a way to inject var
instead of '{{var}}'. Less typing, and typically when you need { or } there
will be whitespace around them. If you really do need to write {something}
literally, {"something"} seems appropriate, or perhaps {"""something"""}
could be allowed for things with lots of quotes?

I miss having redoak in a functional state!